+++ src/components/common/post/dropdown/Dropdown.tsx
@@ -1,7 +1,7 @@
 import { IOption } from '@/types/request';
 import React, { FC, useState } from 'react';
 import Select, { components } from 'react-select';
-import styled from 'styled-components';
+import * as S from './Dropdown.styled';
 
 const customStyles = {
 	menu: (provided: any, state: any) => ({
@@ -64,7 +64,7 @@ const DropdownIndicator = (props: any) => {
 	return (
 		components.DropdownIndicator && (
 			<components.DropdownIndicator {...props}>
-				<BouncingArrow>▼</BouncingArrow>
+				<S.BouncingArrow>▼</S.BouncingArrow>
 			</components.DropdownIndicator>
 		)
 	);
@@ -84,7 +84,7 @@ const Dropdown: React.FC<PropsType> = ({ options, generation, setGeneration }) =
 	};
 
 	return (
-		<Wrapper>
+		<S.Wrapper>
 			<Select
 				isMulti
 				isClearable={false}
@@ -97,35 +97,8 @@ const Dropdown: React.FC<PropsType> = ({ options, generation, setGeneration }) =
 				noOptionsMessage={() => '모든 옵션이 선택되었습니다.'}
 				isSearchable={false}
 			/>
-		</Wrapper>
+		</S.Wrapper>
 	);
 };
 
 export default Dropdown;
-
-const Wrapper = styled.div`
-	font-size: 1.5rem;
-	font-family: Pretendard;
-`;
-
-const BouncingArrow = styled.div`
-	padding: 2px;
-	font-size: 1.8rem;
-	color: blue;
-	animation: bounce 2s infinite;
-	@keyframes bounce {
-		0%,
-		20%,
-		50%,
-		80%,
-		100% {
-			transform: translateY(0);
-		}
-		40% {
-			transform: translateY(10px);
-		}
-		60% {
-			transform: translateY(5px);
-		}
-	}
-`;
